move zoom to different file
